Updated parsing of total energies
The `final` tag previously used for the last ionic step is replaced by `last` in order
to release `final` for use for e.g. the total energy that is present after the electronic steps
are complated (includes potential corrections). In addition, we remove the two entries (`initial` and `final` that was previously present for static runs). For static runs, only one entry is now returned. This means that now, if one needs both the `initial` and `last` property, the only way to get this is by parsing all the ionic entries by enabling `status=all`. In addition it is now possible to parse multiple total energy types. Tests have also been modified accordingly.